I encountered some weirdness with the newstat code previously pushed. With the "old" phdf5/1.10.4 install on Frontera (which is default when you module load phdf5) I get a weird segfault somewhere in lofs-write *if compiled with optimization -O3*. What is especially weird is this only occurs with the newstats code, but the segfault happens before the new code!! So who knows what the fuck is going on. Compiling with -O2 does not result in segfault.
Instead, loading phdf5/1.12.0 and compiling with -O3 works w/out issue.
But I wanted to make a note of this weirdness in case it is ever
encountered again, and because changes to my personal Makefile are not
recorded in git.

In this push, I have #ifdef'ed out the original code that dumps the
stats after the model is done completely running, because all the values
are zero since I haven't used different variable names (that are
overwritten every restart save cycle). I'll probably yank all the code
but for now I've left in place in case I ever decide to go back to that.