Added a bunch of configuration files for strong scaling tests. Using AACP sounding and domain size, but set up for 100 meters 50 meters 25 meters grid spacing (isotropic).
We add 8X the number of gridpoints going from 100m to 50m, and 16X the number of calculations (because we halve the time step)
We add 4X the number of nodes going from 100m to 50m

We add 8X the number of gridpoints going from 50m to 25m, and 16X the number of calculations (because we halve the time step)
We add 4X the number of nodes going from 50m to 25m

Keep these scaling issues in mind when evaluating how the model
performs. CM1 does not do a parallel decomposition in the vertical, so
each MPI rank carries the full vertical extent of the domain.